+++ source/core/mortar_coupling_manager.cc
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025 The Lethe Authors
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception OR LGPL-2.1-or-later
+
+#include "core/mortar_coupling_manager.h"
+
+template class MortarManager<2>;
+template class CouplingOperator<2>;